// src/interfaces/Images.ts
// Strictly used for cloudinary image plugin
export interface ImageFormat {
  name: string;
  hash: string;
  ext: string;
  mime: string;
  path: string | null;
  width: number;
  height: number;
  size: number;
  url: string;
  provider_metadata: {
    public_id: string;
    resource_type: string;
  };
}

export type ImageFormats = {
  thumbnail: ImageFormat;
  small: ImageFormat;
  medium: ImageFormat;
};

export interface Image {
  data: {
    id: number;
    attributes: {
      name: string;
      alternativeText: string | null;
      caption: string | null;
      width: number;
      height: number;
      formats: ImageFormats;
      hash: string;
      ext: string;
      mime: string;
      size: number;
      url: string;
      previewUrl: string | null;
      provider: string;
      provider_metadata: {
        public_id: string;
        resource_type: string;
      };
      createdAt: string;
      updatedAt: string;
    };
  };
}

export default interface Images {
  data: Image[];
}